graphql: Move ObjectCondition to graphql::schema::ast
