+++ graphql/src/lib.rs
@@ -30,7 +30,7 @@ pub mod prelude {
     pub use super::introspection::{introspection_schema, IntrospectionResolver};
     pub use super::query::{execute_query, ext::BlockConstraint, QueryExecutionOptions};
     pub use super::schema::{api_schema, APISchemaError};
-    pub use super::store::{build_query, StoreResolver};
+    pub use super::store::StoreResolver;
     pub use super::subscription::SubscriptionExecutionOptions;
     pub use super::values::MaybeCoercible;
