+++ graphql/src/store/mod.rs
@@ -2,5 +2,5 @@ mod prefetch;
 mod query;
 mod resolver;
 
-pub use self::query::{build_query, parse_subgraph_id};
+pub use self::query::parse_subgraph_id;
 pub use self::resolver::StoreResolver;
